+++ packages/camera/camera_android_camerax/test/android_camera_camerax_test.dart
@@ -50,7 +50,7 @@ import 'package:camera_android_camerax/src/zoom_state.dart';
 import 'package:camera_platform_interface/camera_platform_interface.dart';
 import 'package:flutter/services.dart'
     show DeviceOrientation, PlatformException, Uint8List;
-import 'package:flutter/widgets.dart' show BuildContext, Size, Texture, Widget;
+import 'package:flutter/widgets.dart' show BuildContext, Size;
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
@@ -201,6 +201,8 @@ void main() {
             Future<Camera2CameraInfo>.value(MockCamera2CameraInfo()),
         getUiOrientation: () =>
             Future<DeviceOrientation>.value(DeviceOrientation.portraitUp),
+        previewSurfaceProducerHandlesCropAndRotation: (_) =>
+            Future<bool>.value(false),
       );
 
   /// CameraXProxy for testing exposure and focus related controls.
@@ -919,7 +921,9 @@ void main() {
     expect(camera.recorder!.qualitySelector, isNull);
   });
 
-  test('createCamera sets sensor orientation as expected', () async {
+  test(
+      'createCamera sets sensor orientation, handlesCropAndRotation, initialDeviceOrientation as expected',
+      () async {
     final AndroidCameraCameraX camera = AndroidCameraCameraX();
     const CameraLensDirection testLensDirection = CameraLensDirection.back;
     const int testSensorOrientation = 270;
@@ -929,6 +933,7 @@ void main() {
         sensorOrientation: testSensorOrientation);
     const bool enableAudio = true;
     const ResolutionPreset testResolutionPreset = ResolutionPreset.veryHigh;
+    const bool testHandlesCropAndRotation = true;
     const DeviceOrientation testUiOrientation = DeviceOrientation.portraitDown;
 
     // Mock/Detached objects for (typically attached) objects created by
@@ -948,6 +953,8 @@ void main() {
         getProxyForTestingResolutionPreset(mockProcessCameraProvider);
     camera.proxy.getSensorOrientation =
         (_) async => Future<int>.value(testSensorOrientation);
+    camera.proxy.previewSurfaceProducerHandlesCropAndRotation =
+        (_) async => Future<bool>.value(testHandlesCropAndRotation);
     camera.proxy.getUiOrientation =
         () async => Future<DeviceOrientation>.value(testUiOrientation);
 
@@ -960,7 +967,7 @@ void main() {
     await camera.createCamera(testCameraDescription, testResolutionPreset,
         enableAudio: enableAudio);
 
-    expect(camera.sensorOrientation, testSensorOrientation);
+    expect(camera.sensorOrientationDegrees, testSensorOrientation);
   });
 
   test(
@@ -1296,6 +1303,8 @@ void main() {
     expect(camera.cameraControl, equals(mockCameraControl));
   });
 
+  // Further `buildPreview` testing concerning the Widget that it returns is
+  // located in preview_rotation_test.dart.
   test(
       'buildPreview throws an exception if the preview is not bound to the lifecycle',
       () async {
@@ -1308,22 +1317,6 @@ void main() {
 
     expect(
         () => camera.buildPreview(cameraId), throwsA(isA<CameraException>()));
-  });
-
-  test(
-      'buildPreview returns a Texture once the preview is bound to the lifecycle if it is backed by a SurfaceTexture',
-      () async {
-    final AndroidCameraCameraX camera = AndroidCameraCameraX();
-    const int cameraId = 37;
-
-    // Tell camera that createCamera has been called and thus, preview has been
-    // bound to the lifecycle of the camera.
-    camera.previewInitiallyBound = true;
-
-    final Widget widget = camera.buildPreview(cameraId);
-
-    expect(widget is Texture, isTrue);
-    expect((widget as Texture).textureId, cameraId);
   });
 
   group('video recording', () {