+++ packages/google_maps_flutter/google_maps_flutter_platform_interface/CHANGELOG.md
@@ -1,6 +1,11 @@
+## 2.11.1
+
+* Updates READMEs and API docs.
+
 ## 2.11.0
 
 * Adds support for animating the camera with a duration.
+* Updates READMEs and API docs.
 
 ## 2.10.0