+++ packages/video_player/video_player_android/CHANGELOG.md
@@ -1,5 +1,10 @@
-## NEXT
+## 2.8.1
 
+* Updates compileSdk 34 to flutter.compileSdkVersion.
+
+## 2.8.0
+
+* Adds support for platform views as an optional way of displaying a video.
 * Suppresses deprecation and removal warnings for
   `TextureRegistry.SurfaceProducer.onSurfaceDestroyed`.